+++ DataModels/VM/LogBook/LogBookInstrumentVM.cs
@@ -7,9 +7,9 @@ namespace DataModels.VM.LogBook
     {
         public long Id { get; set; }
         public long LogBookId { get; set; }
-        public short ActualInstrument { get; set; }
-        public short SimulatedInstrument { get; set; }
-        public short Holds { get; set; }
+        public double ActualInstrument { get; set; }
+        public double SimulatedInstrument { get; set; }
+        public double Holds { get; set; }
 
         public List<DropDownSmallValues> Approaches { get; set; } = new ();
         public List<LogBookInstrumentApproachVM> LogBookInstrumentApproachesList { get; set; } = new();
