+++ DataModels/VM/LogBook/LogBookTrainingDetailVM.cs
@@ -6,10 +6,10 @@ namespace DataModels.VM.LogBook
     {
         public long Id { get; set; }
         public long LogBookId { get; set; }
-        public short DualGiven { get; set; }
-        public short DualReceived { get; set; }
-        public short SimulatedFlight { get; set; }
-        public short GroundTraining { get; set; }
+        public double DualGiven { get; set; }
+        public double DualReceived { get; set; }
+        public double SimulatedFlight { get; set; }
+        public double GroundTraining { get; set; }
         public bool IsFlightReview { get; set; }
         public bool IsIPC { get; set; }
         public bool IsCheckRide { get; set; }
